Solution
Exercise 13: each print command usually prints out a line of its own, complete with a change of line at the end. However, if the print command is given an additional argument end = "", it will not print a line change. Fix this program so that the entire calculation, complete with result, is printed out on a single line. Do not change the number of print commands used.