Solution
Exercise 27: This program calculates the end of year bonus a customer receives on their loyalty card. The bonus is calculated with the following formula:

- If there are less than a hundred points on the card, the bonus is 10 %
- In any other case the bonus is 15 %

But there is a problem with the program, so with some inputs it doesn't work quite right. Please fix the program so that there is always either a 10 % or a 15 % bonus, but never both.