Solution
Exercise 48: write a program which keeps asking the user for a PIN code until they type in the correct one, which is 4321. The program should then print out the number of times the user tried different codes. If the user gets it right on the first try, the program should print out something a bit different.