Solution
Exercise 51: write a program which asks the user for integer numbers. The program should keep asking for numbers until the user types in zero. After reading in the numbers the program should print out how many numbers were typed in. The zero at the end should not be included in the count. The program should also print out the sum of all the numbers typed in. The zero at the end should not be included in the calculation. The program should also print out the mean of the numbers. The zero at the end should not be included in the calculation. You may assume the user will always type in at least one valid non-zero number. The program should also print out statistics on how many of the numbers were positive and how many were negative. The zero at the end should not be included in the calculation.