Solution
Exercise 73: write a program which finds the second occurrence of a substring. If there is no second (or first) occurrence, the program should print out a message accordingly.
In this exercise the occurrences cannot overlap. For example, in the string aaaa the second occurrence of the substring aa is at index 2.